tests/testsuite: move UpdateKubeVirtConfigValue to the config package
